Removed bone_data.order from CCDIK
To be honest I don't understand why Godot CCDIK have an "index"
in the joint data. I thought that was to select the order but It's
better to remove than writing something that doesn't make sense.

(my guess is that it needed to get the right bone from the skeleton but
this is not needed in our case)